+++ core/src/main/java/org/apache/calcite/sql/SqlDialectFactoryImpl.java
@@ -16,8 +16,6 @@
  */
 package org.apache.calcite.sql;
 
-import org.apache.calcite.avatica.util.Casing;
-import org.apache.calcite.config.NullCollation;
 import org.apache.calcite.sql.dialect.AccessSqlDialect;
 import org.apache.calcite.sql.dialect.AnsiSqlDialect;
 import org.apache.calcite.sql.dialect.BigQuerySqlDialect;
@@ -68,35 +66,18 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
       JethroDataSqlDialect.createCache();
 
   @Override public SqlDialect create(DatabaseMetaData databaseMetaData) {
-    String databaseProductName;
-    int databaseMajorVersion;
-    int databaseMinorVersion;
-    String databaseVersion;
+    SqlDialect.Context c = SqlDialects.createContext(databaseMetaData);
+
+    String databaseProductName = c.databaseProductName();
     try {
-      databaseProductName = databaseMetaData.getDatabaseProductName();
-      databaseMajorVersion = databaseMetaData.getDatabaseMajorVersion();
-      databaseMinorVersion = databaseMetaData.getDatabaseMinorVersion();
-      databaseVersion = databaseMetaData.getDatabaseProductVersion();
+      if (databaseProductName == null) {
+        databaseProductName = databaseMetaData.getDatabaseProductName();
+      }
     } catch (SQLException e) {
       throw new RuntimeException("while detecting database product", e);
     }
     final String upperProductName =
         databaseProductName.toUpperCase(Locale.ROOT).trim();
-    final String quoteString = getIdentifierQuoteString(databaseMetaData);
-    final NullCollation nullCollation = getNullCollation(databaseMetaData);
-    final Casing unquotedCasing = getCasing(databaseMetaData, false);
-    final Casing quotedCasing = getCasing(databaseMetaData, true);
-    final boolean caseSensitive = isCaseSensitive(databaseMetaData);
-    final SqlDialect.Context c = SqlDialect.EMPTY_CONTEXT
-        .withDatabaseProductName(databaseProductName)
-        .withDatabaseMajorVersion(databaseMajorVersion)
-        .withDatabaseMinorVersion(databaseMinorVersion)
-        .withDatabaseVersion(databaseVersion)
-        .withIdentifierQuoteString(quoteString)
-        .withUnquotedCasing(unquotedCasing)
-        .withQuotedCasing(quotedCasing)
-        .withCaseSensitive(caseSensitive)
-        .withNullCollation(nullCollation);
     switch (upperProductName) {
     case "ACCESS":
       return new AccessSqlDialect(c);
@@ -175,73 +156,6 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
     }
   }
 
-  private static Casing getCasing(DatabaseMetaData databaseMetaData, boolean quoted) {
-    try {
-      if (quoted
-          ? databaseMetaData.storesUpperCaseQuotedIdentifiers()
-          : databaseMetaData.storesUpperCaseIdentifiers()) {
-        return Casing.TO_UPPER;
-      } else if (quoted
-          ? databaseMetaData.storesLowerCaseQuotedIdentifiers()
-          : databaseMetaData.storesLowerCaseIdentifiers()) {
-        return Casing.TO_LOWER;
-      } else if (quoted
-          ? (databaseMetaData.storesMixedCaseQuotedIdentifiers()
-              || databaseMetaData.supportsMixedCaseQuotedIdentifiers())
-          : (databaseMetaData.storesMixedCaseIdentifiers()
-              || databaseMetaData.supportsMixedCaseIdentifiers())) {
-        return Casing.UNCHANGED;
-      } else {
-        return Casing.UNCHANGED;
-      }
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce casing", e);
-    }
-  }
-
-  private static boolean isCaseSensitive(DatabaseMetaData databaseMetaData) {
-    try {
-      return databaseMetaData.supportsMixedCaseIdentifiers()
-          || databaseMetaData.supportsMixedCaseQuotedIdentifiers();
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce case-sensitivity", e);
-    }
-  }
-
-  private static NullCollation getNullCollation(DatabaseMetaData databaseMetaData) {
-    try {
-      if (databaseMetaData.nullsAreSortedAtEnd()) {
-        return NullCollation.LAST;
-      } else if (databaseMetaData.nullsAreSortedAtStart()) {
-        return NullCollation.FIRST;
-      } else if (databaseMetaData.nullsAreSortedLow()) {
-        return NullCollation.LOW;
-      } else if (databaseMetaData.nullsAreSortedHigh()) {
-        return NullCollation.HIGH;
-      } else if (isBigQuery(databaseMetaData)) {
-        return NullCollation.LOW;
-      } else {
-        throw new IllegalArgumentException("cannot deduce null collation");
-      }
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce null collation", e);
-    }
-  }
-
-  private static boolean isBigQuery(DatabaseMetaData databaseMetaData)
-      throws SQLException {
-    return databaseMetaData.getDatabaseProductName()
-        .equals("Google Big Query");
-  }
-
-  private static String getIdentifierQuoteString(DatabaseMetaData databaseMetaData) {
-    try {
-      return databaseMetaData.getIdentifierQuoteString();
-    } catch (SQLException e) {
-      throw new IllegalArgumentException("cannot deduce identifier quote string", e);
-    }
-  }
-
   /** Returns a basic dialect for a given product, or null if none is known. */
   static @Nullable SqlDialect simple(SqlDialect.DatabaseProduct databaseProduct) {
     switch (databaseProduct) {
@@ -313,5 +227,4 @@ public class SqlDialectFactoryImpl implements SqlDialectFactory {
       return null;
     }
   }
-
 }
